:recycle: Modularizado de BBII, muchos componentes en 1 archivo
